chore: remove gemini folder
