+++ src/main/java/com/gdg/kkia/chatbot/repository/ChatbotResponseRepository.java
@@ -1,6 +1,6 @@
 package com.gdg.kkia.chatbot.repository;
 
-import com.gdg.kkia.chatbot.entity.ChatbotResponse;
+import com.gdg.kkia.chatbot.dto.ChatbotResponse;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
